card margin spacing, renamed location_setup to location_change
